Rename commands module
