+++ src/bin/brace/commands/web/run.rs
@@ -37,7 +37,7 @@ pub fn exec(shell: &mut Shell, matches: &ArgMatches) -> ExecResult {
                 let config = overload_file(file, config, shell, matches)?;
 
                 shell.info(format!("Using configuration file: {}", file))?;
-                brace::commands::web::run(config);
+                brace::app::web::run(config);
 
                 Ok(())
             }
@@ -50,7 +50,7 @@ pub fn exec(shell: &mut Shell, matches: &ArgMatches) -> ExecResult {
             let config = overload_default(shell, matches)?;
 
             shell.warn("No configuration file specified")?;
-            brace::commands::web::run(config);
+            brace::app::web::run(config);
 
             Ok(())
         }
