+++ src/lib/app/init.rs
