+++ tests/web.rs
@@ -51,7 +51,7 @@ fn test_web_server_with_config() {
     let temp_dir = TempDir::new().unwrap();
     let temp_path = temp_dir.path();
 
-    brace::commands::init::init(temp_path.to_str().unwrap()).unwrap();
+    brace::app::init::init(temp_path.to_str().unwrap()).unwrap();
 
     let conf_path = temp_path.join("Config.toml");
     let mut conf_file = OpenOptions::new()
